fix users reactivity
