+++ src/lib/stores/webrtc-sync.svelte.ts
@@ -3,13 +3,33 @@ import * as Y from 'yjs';
 import { env } from '$env/dynamic/public';
 import { page } from '$app/state';
 import { replaceState } from '$app/navigation';
+import { browser } from '$app/environment';
+import { ACTIVE_USER_KEY } from '$lib/constants';
+import type { User } from '$lib/types';
+import randomColor from 'randomcolor';
 
 export class WebrtcSync {
 	isSyncing = $state<boolean>(false);
 	provider = $state<WebrtcProvider | null>(null);
+	activeUser = $state<User>({
+		name: '',
+		color: ''
+	});
 	private queryParamName = 'sync';
 
-	constructor(id: string, ydoc: Y.Doc) {
+	constructor(id: string, ydoc: Y.Doc, username: string = 'Anonymous') {
+		// Initialize active user
+		this.activeUser = {
+			name: username,
+			color: randomColor()
+		};
+
+		if (browser) {
+			const item = localStorage.getItem(ACTIVE_USER_KEY);
+			if (item) this.activeUser = this.deserialize(item);
+		}
+
+		// Initialize sync state from URL
 		const urlParams = new URLSearchParams(window.location.search);
 		const syncParam = urlParams.get(this.queryParamName);
 		if (syncParam === 'true') {
@@ -39,6 +59,16 @@ export class WebrtcSync {
 				this.provider = null;
 			}
 		});
+
+		$effect(() => {
+			localStorage.setItem(ACTIVE_USER_KEY, this.serialize(this.activeUser));
+			if (this.provider) {
+				this.provider.awareness.setLocalStateField('user', {
+					name: this.activeUser.name,
+					color: this.activeUser.color
+				});
+			}
+		});
 	}
 
 	toggleSync = () => {
@@ -52,4 +82,12 @@ export class WebrtcSync {
 		}
 		replaceState(url, {});
 	};
+
+	serialize(value: User): string {
+		return JSON.stringify(value);
+	}
+
+	deserialize(item: string): User {
+		return JSON.parse(item);
+	}
 }
